test/web-platform-tests/to-upstream/html/semantics/scripting-1/the-script-element/execution-timing/async-script-depending-on-previous-sync-dont-upstream.html

<!DOCTYPE html>
<html>
    <head>
        <title>An async script executes after a previous sync script is executed</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="This test ensures an async script executes after a previous sync script is executed" />
        <link rel="author" title="Jesus David Garcia Gomez" href="mailto:jdgarcia@outlook.com" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/#attr-script-async"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script type="text/javascript">
            var t = async_test("An async script executes after a previous sync script is executed");

            function timeout()
            {
                t.step(function(){ assert_equals(document.getElementById("testresult").innerHTML, "run")});
                t.done();
            }

            var timer = setTimeout(timeout, 4000);

            function log(text)
            {
                var textNode = document.createTextNode(text);
                document.getElementById("testresult").appendChild(textNode);
            }
        </script>

        <span id="testresult"></span>

        <script src="sync.py?sec=3"></script>
        <script src="async.js" async></script>
    </body>
</html>
